requirements.txt
# GoAI Sovereign Platform v1 - Dependencies

# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Data Validation
pydantic==2.5.2

# File Upload Support
python-multipart==0.0.6

# Vector Store
faiss-cpu==1.7.4
numpy==1.26.2

# Configuration
PyYAML==6.0.1

# HTTP Client (for LLM providers)
httpx==0.25.2
aiohttp==3.9.1

# Environment Variables
python-dotenv==1.0.0

# ============================================
# DATABASE
# ============================================
sqlalchemy==2.0.23
psycopg2-binary==2.9.9  # PostgreSQL driver
alembic==1.12.1  # Migrations
asyncpg==0.29.0  # Async PostgreSQL

# ============================================
# CACHING & SESSIONS
# ============================================
redis==5.0.1
hiredis==2.3.2  # Redis speedup

# ============================================
# SECURITY
# ============================================
passlib[bcrypt]==1.7.4  # Password hashing
python-jose[cryptography]==3.3.0  # JWT tokens

# ============================================
# ASYNC SUPPORT
# ============================================
asyncio-throttle==1.0.2

# ============================================
# DOCUMENT PARSING (Page-Aware)
# ============================================
pypdf==3.17.0  # PDF extraction with page tracking
python-docx==1.1.0  # DOCX extraction
openpyxl==3.1.2  # Excel support
chardet==5.2.0  # Text encoding detection

# ============================================
# OPTIONAL (uncomment as needed)
# ============================================
# spacy==3.7.2  # Advanced NLP
# sentence-transformers==2.2.2  # Local embeddings
# pgvector==0.2.4  # PostgreSQL vector extension
